Moved focus setters & updated useEffects to set focus on h1 on each page
