+++ src/applications/combined-debt-portal/debt-letters/containers/DebtLettersSummary.jsx
@@ -1,7 +1,11 @@
 import React, { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
-import scrollToTop from 'platform/utilities/ui/scrollToTop';
+import {
+  setPageFocus,
+  ALERT_TYPES,
+  APP_TYPES,
+} from '../../combined/utils/helpers';
 import HowDoIPay from '../components/HowDoIPay';
 import NeedHelp from '../components/NeedHelp';
 import DebtCardsList from '../components/DebtCardsList';
@@ -10,7 +14,6 @@ import '../sass/debt-letters.scss';
 // TODO: OtherVA Update
 import OtherVADebts from '../../combined/components/OtherVADebts';
 import alertMessage from '../../combined/utils/alert-messages';
-import { ALERT_TYPES, APP_TYPES } from '../../combined/utils/helpers';
 
 const renderAlert = (alertType, statements) => {
   const alertInfo = alertMessage(alertType, APP_TYPES.DEBT);
@@ -69,7 +72,7 @@ const DebtLettersSummary = () => {
   const { statements: mcpStatements, error: mcpError } = mcp;
 
   useEffect(() => {
-    scrollToTop();
+    setPageFocus('h1');
   }, []);
 
   if (isDebtPending || isPendingVBMS || isProfileUpdating) {
