+++ src/applications/combined-debt-portal/debt-letters/utils/page.js
@@ -1,21 +1,3 @@
-export const setFocus = selector => {
-  const el =
-    typeof selector === 'string' ? document.querySelector(selector) : selector;
-  if (el) {
-    el.setAttribute('tabIndex', -1);
-    el.focus();
-  }
-};
-
-export const setPageFocus = (selector = '.va-nav-breadcrumbs') => {
-  const el = document.querySelector(selector);
-  if (el) {
-    setFocus(el);
-  } else {
-    setFocus('#main h1');
-  }
-};
-
 export const currency = new Intl.NumberFormat('en-US', {
   style: 'currency',
   currency: 'USD',
